+++ src/main/java/devmagic/Dto/CartItemDTO.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -12,8 +14,8 @@ public class CartItemDTO {
     private String imageUrl;
     private String productName;
     private int quantity;
-    private double price;
-    private Product product; // Thêm thuộc tính Product
+    private BigDecimal price; // Thay đổi kiểu từ double sang BigDecimal
+    private Product product; // Thuộc tính Product
 
     public Product getProduct() {
         return product;
